chore: rename the JS file and add method description
