+++ frontend/src/components/dashboard/users/ImportModal.vue
@@ -172,7 +172,7 @@ import Papa from "papaparse";
 import { downloadObjectsAs } from "@/assets/utils.js";
 import MoodleOptions from "@/basic/form/MoodleOptions.vue";
 import RoleMappingStep from "@/components/dashboard/users/RoleMappingStep.vue";
-import { buildInitialRoleMappings, getRoleRows, normalizeImportUsers } from "@/components/dashboard/users/moodleRoleMapping.js";
+import { buildInitialRoleMappings, getRoleRows, normalizeImportUsers } from "@/components/dashboard/users/roleMapping.js";
 
 /**
  * Modal for bulk creating users through csv file and Moodle API
